chore: rearranged babydegen file structure
