+++ components/Services/BabyDegenPage/index.jsx
@@ -1,0 +1,17 @@
+import { FurtherResources } from './FurtherResources';
+import { Hero } from './Hero';
+import { HowItWorks } from './HowItWorks';
+import { MeetTheBabyDegens } from './MeetTheBabyDegens';
+import { WhatIsBabyDegen } from './WhatIsBabyDegen';
+
+const BabyDegen = () => (
+  <>
+    <Hero />
+    <WhatIsBabyDegen />
+    <MeetTheBabyDegens />
+    <HowItWorks />
+    <FurtherResources />
+  </>
+);
+
+export default BabyDegen;
